cli: support multiple logs from a single policy
Expose the capability to have a policy file with multiple logs to the
user.

We still want to keep the simple execution model of one process
monitoring one log, so we do this by allowing the user to select which
log in the policy to monitor. If they want to monitor multiple/all of
them, they can just spawn one for each.

However, this also means that we now have to deal with concurrent access
to the state file. Avoid any potential headaches by simply having one
file per log and putting them in a new subdirectory. Note that we name
the file after the log pubkey instead of the keyhash (which would be
more in line with the general vibe of the sigsum ecosystem). The
reasoning here is that it should be easy for the user to map from a
"log" statement in the policy file to the state file, so using the same
value is preferable.