cli: move monitor state into sub-object
Move the monitor state below the "monitor" key in the state file while
there are not many deployments that are affected by such a breaking
change.

This allows us to reduce the amount of pain we cause if we ever want to
add auxiliary information to this file in the future.