cli: add new hook type for augmenting leaves
sigsum leaves by design don't contain any content that is meaningfully
user-controllable. While this prevents injection of unwanted data into
the log, it also means that a monitor can a priori only report that a
key *has* been used, but not *how*.

However, this would be useful to help the user decide if a signing event
was intentional or not. As it stands, the user must manually do this by
performing a reverse lookup of the checksum in some application-specific
manner.

Add a new hook type ("leaf_info") that can be used to automate this:
Hooks of this type run before the "event" hooks and get the same
information an event hooks gets. They can use that information to
retrieve additional information about the signing event. If a leaf
information hook is successful and produces output on stdout, it gets
added to an environment variable that event hooks can then use to
augment the notifications they send out.