add log staleness monitoring
We're currently sending out events whenever a new matching leaf appears.
However, the expectation also is that if no events are emitted, no new
matching leaves have appeared in the log.

For this, we need to make sure that the log is still processing new
leaves and serving the latest consensus reality to us.

Thus, we also need to add alerting when the log *isn't* doing that (so
that absence of such alerts assures us the log is behaving as expected).

Add a new "log_health" hook type and allow the user to specify a maximum
age of the tree head before this hook is triggered.