sigsum: calculate quorum timestamp
Currently, we are not making any use of the timestamps that are part of
the witness cosignatures. However, they contain valuable information
since they assert that:

 - The cosigned tree head was the most recent one observed by that
   witness at the time of cosigning
 - The log has been consistent up to this point in time

What makes this even more valuable is that in Sigsum, the log
re-requests cosignatures regularly even if the tree didn't grow. So we
can use the *absence* of fresh cosignatures as a sign that we are seeing
a stale version of the tree head for whatever reason.

The only question is how to meaningfully distill the timestamps from the
various witnesses into a single one, since it is conceivable that
individual witnesses might have drifting clocks or might be
intermittently unresponsive when the log is refreshing the cosignatures,
for example. In particular, we should only consider cosignatures from
witnesses that we trust (since a malicious log could always add bogus
cosignatures).

Here, we conceptually use the following approach:

 - If no quorum is required ("quorum none" in the trust policy), we use
   the latest cosignature timestamp

 - Otherwise, we start with the cosignature with the most recent
   timestamp, and check if the quorum policy is satisfied. If not, we
   add the next older cosignature and so on until we either have reached
   a quorum or exhausted all cosignatures.

   The last timestamp we've had add to reach a quorum is the quorum
   timestamp.

The algorithm in this commit is a little less optimized than it could be
(we are re-evaluating the whole policy during each iteration, even if
parts of it are already satisfied), but this makes the whole thing much
easier to understand and isn't a big deal in practice, since the
timestamps are integers and we add all cosignatures with the same
timestamp at once. Since normally, all the witness timestamps are more
or less the same, this means we only need to do one or two iterations of
the outer loop.