add log mirroring
Add a tool for mirroring a Sigsum log into the tlog-tiles format.

This is somewhat unrelated to the sigmon core business and at some point
we ought to factor out the common Sigsum API logic into a library so
that we can split sigmon and tools like this into their separate
projects. But this project is still evolving and that would add
unnecessary friction so let's keep everything together for the time
being.