move quorum handling to toplevel
Currently, SigsumLogAPI.get_tree_head() checks that a tree head
satisfies the quorum.

However, this mixes concerns somewhat and also is less flexible than we
would like, for example we might want to check a tree head against
multiple policies and there might not be a distinguished one. Or we
might want to apply more complex rules on how to aggregate witness
cosignatures in the context of a monitor.

Drop the quorum check from SigsumLogAPI and move it to the top level
loop so that the behavior doesn't change for now.